backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-docx==1.1.0
python-multipart==0.0.9
pydantic==2.5.2
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0 
anthropic
python-dotenv
